+++ spec/action/attachable/axn_spec.rb
@@ -36,9 +36,9 @@ RSpec.describe Action do
         expect(client.foo).not_to be_ok
 
         # The automatic logging will log before and after execution, so we need to expect multiple calls
-        expect(Action.config.logger).to receive(:debug).with(/About to execute with: {expected: true, arg: 123}/).ordered
-        expect(Action.config.logger).to receive(:info).with(/got expected=true, arg=123/).ordered
-        expect(Action.config.logger).to receive(:debug).with(/Execution completed \(with outcome: success\)/).ordered
+        expect(Axn.config.logger).to receive(:debug).with(/About to execute with: {expected: true, arg: 123}/).ordered
+        expect(Axn.config.logger).to receive(:info).with(/got expected=true, arg=123/).ordered
+        expect(Axn.config.logger).to receive(:debug).with(/Execution completed \(with outcome: success\)/).ordered
         expect(client.foo(expected: true, arg: 123)).to be_ok
       end